feat: refactor torrent service by modularizing components and enhancing stream handling
